LPD-85450 Consolidate create-object-field tests into objectField.spec.ts
- Move the 5 new tests from tests/object-web/main/createObjectFields.spec.ts
  into a nested 'Create Object Fields' describe under the existing
  'Manage objectFields through Objects Admin UI' block in
  tests/object-web/field/objectField.spec.ts (sorted alphabetically).
- Drop the new openObjectFieldByDropdownAction helper; the existing
  openObjectField method already handles the post-save re-open scenario,
  so the two callsites in the migrated update test now call openObjectField.
